requirements.txt
selenium
tqdm
psycopg2-binary
python-dotenv
python-dotenv>=1.0.0
